docker/root/requirements.txt
psutil
pycountry
requests
parse-torrent-title
rapidfuzz
thefuzz
colorlog
pyinotify
python-dotenv
websocket-client
mysql-connector-python
msgspec
starlette
uvicorn[standard]